Move part type to Utils candidates
